+++ packages/react/src/components/DetailsList/DetailsList.test.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import * as renderer from 'react-test-renderer';
 import { render, fireEvent, act } from '@testing-library/react';
 
 import { EventGroup, KeyCodes, resetIds } from '../../Utilities';
@@ -9,8 +8,6 @@ import { DetailsHeader } from './DetailsHeader';
 import { DetailsList } from './DetailsList';
 import { DetailsListBase } from './DetailsList.base';
 import { CheckboxVisibility, DetailsListLayoutMode } from './DetailsList.types';
-import { DetailsRow } from './DetailsRow';
-import { DetailsRowCheck } from './DetailsRowCheck';
 import type { IDragDropEvents } from '../../DragDrop';
 import type { IGroup } from '../../GroupedList';
 import type { IRenderFunction } from '../../Utilities';
@@ -89,7 +86,7 @@ describe('DetailsList', () => {
   });
 
   it('renders List correctly', () => {
-    const component = renderer.create(
+    const { container } = render(
       <DetailsList
         items={mockData(5)}
         onRenderRow={() => null}
@@ -97,12 +94,11 @@ describe('DetailsList', () => {
         onShouldVirtualize={() => false}
       />,
     );
-    const tree = component.toJSON();
-    expect(tree).toMatchSnapshot();
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   it('renders List correctly with onRenderDivider props', () => {
-    const component = renderer.create(
+    const { container } = render(
       <DetailsList
         items={mockData(5)}
         columns={mockData(5, true)}
@@ -111,12 +107,11 @@ describe('DetailsList', () => {
         onShouldVirtualize={() => false}
       />,
     );
-    const tree = component.toJSON();
-    expect(tree).toMatchSnapshot();
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   it('renders List with custom icon as column divider', () => {
-    const component = renderer.create(
+    const { container } = render(
       <DetailsList
         items={mockData(5)}
         columns={mockData(5, true, true)}
@@ -125,12 +120,11 @@ describe('DetailsList', () => {
         onShouldVirtualize={() => false}
       />,
     );
-    const tree = component.toJSON();
-    expect(tree).toMatchSnapshot();
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   it('renders List in fixed constrained layout correctly', () => {
-    const component = renderer.create(
+    const { container } = render(
       <DetailsList
         items={mockData(5)}
         onRenderRow={() => null}
@@ -139,8 +133,7 @@ describe('DetailsList', () => {
         onShouldVirtualize={() => false}
       />,
     );
-    const tree = component.toJSON();
-    expect(tree).toMatchSnapshot();
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   it('renders a single proportional column with correct width', () => {
@@ -301,7 +294,7 @@ describe('DetailsList', () => {
   });
 
   it('renders List in compact mode correctly', () => {
-    const component = renderer.create(
+    const { container } = render(
       <DetailsList
         items={mockData(5)}
         onRenderRow={() => null}
@@ -310,12 +303,11 @@ describe('DetailsList', () => {
         onShouldVirtualize={() => false}
       />,
     );
-    const tree = component.toJSON();
-    expect(tree).toMatchSnapshot();
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   it('renders List with hidden checkboxes correctly', () => {
-    const component = renderer.create(
+    const { container } = render(
       <DetailsList
         items={mockData(5)}
         skipViewportMeasures={true}
@@ -337,8 +329,7 @@ describe('DetailsList', () => {
         checkboxVisibility={CheckboxVisibility.hidden}
       />,
     );
-    const tree = component.toJSON();
-    expect(tree).toMatchSnapshot();
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   it('focuses row by index', () => {
@@ -806,7 +797,7 @@ describe('DetailsList', () => {
     const onRenderDetailsHeaderMock = jest.fn();
     const onRenderDetailsFooterMock = jest.fn();
 
-    const component = renderer.create(
+    const component = render(
       <DetailsList
         items={mockData(5)}
         onRenderDetailsHeader={onRenderDetailsHeaderMock}
@@ -814,8 +805,8 @@ describe('DetailsList', () => {
       />,
     );
 
-    const grid = component.root.findByProps({ role: 'grid' });
-    expect(grid.props[`aria-rowcount`]).toEqual(7);
+    const grid = component.getByRole('grid');
+    expect(grid).toHaveAttribute('aria-rowcount', '7');
   });
 
   it('invokes onRenderColumnHeaderTooltip to customize DetailsColumn tooltip rendering when provided', () => {
@@ -941,7 +932,7 @@ describe('DetailsList', () => {
       return <div>{groupDividerProps.group?.name}</div>;
     };
 
-    const component = renderer.create(
+    const { container, rerender } = render(
       <DetailsList
         onRenderDetailsHeader={onRenderDetailsHeader}
         onRenderRow={onRenderRow}
@@ -953,11 +944,11 @@ describe('DetailsList', () => {
       />,
     );
 
-    expect(component.toJSON()).toMatchSnapshot();
+    expect(container.firstChild).toMatchSnapshot();
 
     // New items, same groups
 
-    component.update(
+    rerender(
       <DetailsList
         onRenderDetailsHeader={onRenderDetailsHeader}
         onRenderRow={onRenderRow}
@@ -969,11 +960,11 @@ describe('DetailsList', () => {
       />,
     );
 
-    expect(component.toJSON()).toMatchSnapshot();
+    expect(container.firstChild).toMatchSnapshot();
 
     // Same items, new groups
 
-    component.update(
+    rerender(
       <DetailsList
         onRenderDetailsHeader={onRenderDetailsHeader}
         onRenderRow={onRenderRow}
@@ -985,11 +976,11 @@ describe('DetailsList', () => {
       />,
     );
 
-    expect(component.toJSON()).toMatchSnapshot();
+    expect(container.firstChild).toMatchSnapshot();
 
     // New items, same groups
 
-    component.update(
+    rerender(
       <DetailsList
         onRenderDetailsHeader={onRenderDetailsHeader}
         onRenderRow={onRenderRow}
@@ -1001,7 +992,7 @@ describe('DetailsList', () => {
       />,
     );
 
-    expect(component.toJSON()).toMatchSnapshot();
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   it('handles paged updates to items within groups', () => {
@@ -1048,7 +1039,7 @@ describe('DetailsList', () => {
       return <div>{groupDividerProps.group?.name}</div>;
     };
 
-    const component = renderer.create(
+    const { container, rerender } = render(
       <DetailsList
         onRenderDetailsHeader={onRenderDetailsHeader}
         onRenderRow={onRenderRow}
@@ -1061,11 +1052,11 @@ describe('DetailsList', () => {
       />,
     );
 
-    expect(component.toJSON()).toMatchSnapshot();
+    expect(container.firstChild).toMatchSnapshot();
 
     // New items, same groups
 
-    component.update(
+    rerender(
       <DetailsList
         onRenderDetailsHeader={onRenderDetailsHeader}
         onRenderRow={onRenderRow}
@@ -1078,7 +1069,7 @@ describe('DetailsList', () => {
       />,
     );
 
-    expect(component.toJSON()).toMatchSnapshot();
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   it('returns an element with the correct text based on the second id passed in aria-labelledby', () => {
@@ -1113,7 +1104,7 @@ describe('DetailsList', () => {
   });
 
   it('has an aria-labelledby checkboxId that matches the id of the checkbox', () => {
-    const component = renderer.create(
+    const { container } = render(
       <DetailsList
         items={mockData(5)}
         columns={mockData(5, true)}
@@ -1123,11 +1114,11 @@ describe('DetailsList', () => {
       />,
     );
 
-    const detailsRowSource = component.root.findAllByType(DetailsRow)[0];
-    const detailsRowCheckSource = detailsRowSource.findByType(DetailsRowCheck).props;
-    const checkboxId = detailsRowCheckSource.id;
+    const detailsRowSource = container.querySelector('[data-automationid="DetailsRow"]');
+    const detailsRowCheckSource = detailsRowSource?.querySelector('[data-automationid="DetailsRowCheck"]');
+    const checkboxId = detailsRowCheckSource?.id;
 
-    expect(checkboxId).toEqual(detailsRowCheckSource[`aria-labelledby`].split(' ')[0]);
+    expect(checkboxId).toEqual(detailsRowCheckSource?.getAttribute(`aria-labelledby`)?.split(' ')[0]);
   });
 
   it('names group header checkboxes based on checkButtonGroupAriaLabel', () => {
@@ -1160,12 +1151,12 @@ describe('DetailsList', () => {
   });
 
   it('makes the first row tabbable if isHeaderVisible is false', () => {
-    const component = renderer.create(
+    const { container } = render(
       <DetailsList items={mockData(5)} columns={mockData(5, true)} isHeaderVisible={false} />,
     );
 
-    const firstRow = component.root.findAllByType(DetailsRow)[0];
+    const firstRow = container.querySelector('[data-automationid="DetailsRow"]');
 
-    expect(firstRow.props.focusZoneProps?.tabIndex).toEqual(0);
+    expect(firstRow).toHaveAttribute('tabIndex', '0');
   });
 });