+++ packages/react/src/components/ThemeGenerator/ThemeGenerator.ts
@@ -22,7 +22,7 @@ export class ThemeGenerator {
     isInverted = false,
     isCustomization = false,
     overwriteCustomColor = true,
-  ) {
+  ): void {
     if (!rule.color && rule.value) {
       // not a color rule
       return;
@@ -50,7 +50,7 @@ export class ThemeGenerator {
    * setSlot() can be called before this, but this must be called before getThemeAs*().
    * Does not override colors of rules where isCustomized is true (i.e. doesn't override existing customizations).
    */
-  public static insureSlots(slotRules: IThemeRules, isInverted: boolean) {
+  public static insureSlots(slotRules: IThemeRules, isInverted: boolean): void {
     // Get all the "root" rules, the ones which don't inherit. Then "set" them to trigger updating dependent slots.
     for (const ruleName in slotRules) {
       if (slotRules.hasOwnProperty(ruleName)) {