+++ packages/test-utilities/src/index.ts
@@ -1,7 +1,8 @@
 export { createTestContainer } from './createTestContainer';
 export { getCSSRules } from './getCSSRules';
-export { safeCreate } from './safeCreate';
-export { create } from './reactTestRenderer';
 
-// eslint-disable-next-line @typescript-eslint/no-deprecated
+/* eslint-disable @typescript-eslint/no-deprecated */
+export { safeCreate } from './deprecated/safeCreate';
+export { create } from './deprecated/reactTestRenderer';
 export { safeMount } from './deprecated/safeMount';
+/* eslint-enable @typescript-eslint/no-deprecated */