+++ packages/utilities/src/styled.test.tsx
@@ -1,4 +1,3 @@
-/*  eslint-disable @typescript-eslint/no-deprecated */
 import * as React from 'react';
 import { act, render } from '@testing-library/react';
 import { styled } from './styled';